debug and restructure config
