+++ WeiboSpider/config/__init__.py
@@ -3,6 +3,6 @@
 # @Time    : 2021/7/17 0:21
 # @Function:
 
-from WeiboSpider.spiders.config.UserInfoConfig import UserInfoConfig
+from WeiboSpider.config.UserInfoConfig import UserInfoConfig
 
 __all__ = ['UserInfoConfig']
