+++ WeiboSpider/config/base/__init__.py
@@ -3,6 +3,6 @@
 # @Time    : 2021/7/17 1:28
 # @Function:
 
-from WeiboSpider.spiders.config.base.Config import Config
+from WeiboSpider.config.base.Config import Config
 
 __all__ = ['Config']
